+++ lib/online_migrations/error_messages.rb
@@ -84,6 +84,10 @@ class <%= migration_name %> < <%= migration_parent %>
   end
 end",
 
+      add_column_generated_stored:
+"Adding a stored generated column blocks reads and writes while the entire table is rewritten.
+Add a non-generated column and use callbacks or triggers instead.",
+
       add_column_json:
 "There's no equality operator for the json column type, which can cause errors for
 existing SELECT DISTINCT queries in your application. Use jsonb instead.