+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Add check for `add_unique_key` (Active Record >= 7.1)
 - Add check for `add_column` with stored generated columns
 
 ## 0.7.3 (2023-05-30)