+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+## 0.8.0 (2023-07-24)
+
 - Add check for `change_column_default`
 - Add check for `add_unique_key` (Active Record >= 7.1)
 - Add check for `add_column` with stored generated columns